+++ plone/app/users/browser/personalpreferences.py
@@ -430,7 +430,8 @@ class PasswordAccountPanel(AccountPanelForm):
         # check if password is correct
         current_password = data.get('current_password')
         if current_password:
-            current_password = current_password.encode('ascii', 'ignore')
+            if isinstance(current_password, unicode):
+                current_password = current_password.encode('utf8')
 
             if not membertool.testCurrentPassword(current_password):
                 err_str = _(u"Incorrect value for current password")
@@ -460,6 +461,8 @@ class PasswordAccountPanel(AccountPanelForm):
         membertool = getToolByName(self.context, 'portal_membership')
 
         password = data['new_password']
+        if isinstance(password, unicode):
+            password = password.encode('utf8')
 
         try:
             membertool.setPassword(password, None, REQUEST=self.request)